Rewrote sfd with class fields
Moved signed parameters to field of the class, with appropriate class constructors

+++ AppMeshPOM/Source/terrain.cpp
@@ -5,21 +5,21 @@ double Terrain::Signed(const Vector& vec) const
     // Rotation matrix
     const Mat2 m2 = Mat2(0.8, -0.6, 0.6, 0.8);
 
-    const float l0 = 350.0;
-    const float a0 = 150.0;
+    const double l0 = 350.0;
+    const double a0 = 150.0;
 
     Vector2 vec2(vec[0], vec[2]);
     Vector2  p = vec2 / l0;
-    float a = 0.0;
-    float b = 1.0;
-    for (int i = 0; i < 4; i++)
+    double a = 0.0;
+    double b = 1.0;
+    for (unsigned i = 0; i < 4; i++)
     {
-        float n = Noise(Vector(p[0], p[1], 0.0));
+        double n = Noise(Vector(p[0], p[1], 0.0));
         a += b * n;
         b *= 0.5;
         p = m2 * p * 2.0;
     }
-    const float terrain2D = a + (vec[0] * vec[0]) / 100.0;
+    const double terrain2D = a + (vec[0] * vec[0]) / 100.0;
     return vec[1] - terrain2D;
 }
